requirements.txt
bleak>=0.21.0
paho-mqtt>=1.6.1
flask>=2.3.0
